Fix deploy command
The deploy command had two issues: first, it was publishing the `public`
directory which is the *PRE BUILT* static files.  We actually want to
deploy the `dist` directory.

Second, since we're hosting on github pages we actually need to include
the archive name as part of the base path so vue knows where to load
itself.